+++ CHANGELOG.md
@@ -10,3 +10,7 @@
 ### 1.0.2
 * fixed compilation error when using dynamic frameworks
 * upgraded dependencies
+
+### 1.0.3
+* added background messaging support oon Android
+* added workaround for flutter 1.7